We now properly serialize shows
Note that this commit makes a change to what the default value for the
"loaded_show" QSetting is. It was previously set to a string value, but
it should be a binary representation of a show, which it is now. You
will likely need to delete the registry key for the old setting for this
to work properly. To do this, delete
HKEY_CURRENT_USER\SOFTWARE\University of Cincinnati